Do not equate object pose with search pose
The search location describes any location at the specific area and is not at
all object specific. It makes more sense to create an own, not equated
designator for the object location. This also prevents searching for objects
getting stuck when one not fitting object was found because equation with a
specific pose disabled the costmap.

This commit also fixes a very nasty bug where removal of the at desig property
didn't work anymore.