Fixed handling of setting joint-states without a valid parent
When the joint's parent doesn't have a collision mesh attached, i.e.
if it doesn't have a corresponding rigid body, setting of joint states
didn't work because the system was unable to calculate the new pose of
the child link. Fixed this by implementing the new method
FIND-PARENT-POSE that returns the pose of a joint's parent link. If a
corresponding rigid body is available, the body is used. Otherwise, the
urdf tree is traversed until a rigid body or the tree's root is reached.