Fixed robot-model pose method.
Using the pose method of OBJECT doesn't really work because it returns
the pose of the reference body which corresponds to a collision entry in
the urdf. Though, the collision entry is not always aligned with the
actual link. Instead of using its pose, we need to use the link-pose of
the root body of the urdf.