Changed (and renamed) macro WITH-BULLET-WINDOW-CONTEXT
Different rendering contexts require different ways to render in them.
For rendering in a specific context we provide the macro WITH-GL-CONTEXT
that replaces WITH-BULLET-WINDOW-CONTEXT. The macro calls a method for
running a lambda function in the specific context. That way, each
context can provide its own functionality for running code in the
context.