Created the new package cram_environment_representation.
It contains event handlers and some reasoning functionality to use
bullet_reasoning as an environment model. The event handlers update the
internal world representation when objects are detected or manipulated
or when the robot moved. Predicates are provided to infer if occasions
have been achieved and this package implements the HOLDS predicate.