Added an interface for filtering designator-solutions at resolution time.
Using validation functions implies that filters have to be built up at
instantiation time of a designator. Sometimes, it is necessary to filter
based on information that is only available at resolution time, in
particular when searching for next solutions. Added a functionality that
allows for adding filters at resolution time.