Calculation of put-down pose can ignore the goal orientation now.
For many objects that are rotationally symmetric, the orientation of the
goal pose actually doesn't matter. In that case, we just use the same
orientation of the object in the robot's base when we grasped it for
calculating the put-down pose. This makes sure that the put-down pose is
actually reachable since we already grasped the object.