Use with-copied-world instead of with-stored-world.
It seems like something goes wrong when storing and restoring the world.
In some cases, collision checking and simulation stops working in that
case. Until we find a better fix, we work on a copy instead of storing
and restoring the world state. This breaks the debug display but at
least seems to work.