Extended the handle description by an object designator for handle radius.
The handles on an object are now described by an object designator.
This designator consists of a location designator for the relative handle pose and a radius attribute for the handle itself.
All functions have been refactored to use this new type of description. Grippers are opened and closed appropriately now.