Completed grasp status updating for handle designators
The function `update-handle-grasp-status' now properly defines the gripper in which a handle currently resides. The handle designator as well as the corresponding object designator of the handled object are updated and equated to their previous versions. Also, removed lift-handle-object-with-relative-location from grasp-object-with-handles since it is not part of the grasping. Updates documentation strings accordingly.