Removed the function for updating handle designators for now
Removed the function `update-handle-pregrasp-status' for now since it would be unused. Therefore, designators of object handles are not updated concerning the gripper they are grasped with.
Also removed some unnecessary functions (putdown-with-handles) due to availability of that functionality from other modules (cram_environment_representation).
Also, the radius for handles is now properly implemented in the process module.