Fixed race conditions in pm-execute.
Waiting for the process module status to be :running after sending an
input can lead to a lock if the module processes too quickly. Instead
wait for the input to become NIL.

Added a lock to prevent race conditions when two tasks try to send
something to a process module at the same time.