Fixed a nasty bug that caused an infinite prolog loop.
If we have an object instance, VALID-GRASP is used which binds ?SIDES.
However, if we do not have an object, GRASP is used and ?SIDES is
unbound. That leads to an infinite loop because (MEMBER ?SIDE ?SIDES)
with both variables unbound iterates forever. Changed the code to use
MEMBER only if VALID-GRASP is used. Otherwise use the ARM predicate to
iterate over sides.