Made handle grasp evaluation possible and made code simpler
The handles on objects to grasp are now evaluated by the "effort" the robot has to invest in order to get the grippers into their respective pregrasp-positions for specific handles. The "cheapest" handle in terms of joint-space distance is then chosen. If no gripper side is specified, the gripper with the nearest handle is used on that handle. Also, object-handle-absolute now does not return location designators, but pose-stamped instances since this value is immediately returned into a pose-stamped afterwards. Several doc strings are added, variables with pregrasp- and grasp-offsets during handle grasp are moved to parameters (with their respective doc strings). The overall line lengths are reduced by introducing more newlines.