Added further predicates and action handler for grasping of objects with handles.
There is a new action handler called grasp-slave which should do less reasoning
and uses predetermined arms and grasps to grasp an object.

The new predicate best-grasps shall be a hook for reasoning about how to
grasp an object.

In order to make sure that the original grasping functionalities remain
functional and untouched another fact group that checks for handled
objects was added. In the future grasp-slave and grasp and their
respective fact groups should be merged into one single grasp
action handler and fact group.