Moved functions to the right places, corrected line lengths, extended distance
Functions in the pr2_manipulation_process_module are now moved into the right files (according to their use). Also, corrected line lengths and indentations at several places. Extended the euclidean-distance function to return a distance for each link that is given (before, it just returned the first distance it could get). So for each target-link given, there is a cons cell containing the link name and the resulting distance. Additionally, extended the get-fk function to return the right values as well. Also, overall cleanup.