Implemented optional use of the euclidean-distance measure while grasp eval
The distance for the grippers during grasps can now be measured by either a quadratic joint-space integral or the euclidean distance for one of the robot-links from the current pose to the final grasping pose. The `reaching-length' function now takes an optional parameter `calc-euclidean-distance' with which the euclidean distance measure can be turned on.