Added simplifying function to get euclidean distance for only one link
As the full call to `euclidean-distance' and the interpretation of it's output is pretty cumbersome when only one link is requested (which is the usual case), the simplyfied function `euclidean-distance-for-link' now handles all the data preparation and returns the distance for only one given link. The `reaching-length' function has been adapted appropriately.