Added adapter functions for `ik->trajectory' to overcome type problems
The "(declare (type (or" formulation does not work
here as it does not play well with
`roslisp:with-fields'. `roslisp:with-fields' internally wants to
transform `ik-result' into a list and reads the `type' from
`ik-result', resulting in an error saying "OR
KINEMATICS_MSGS-SRC:GETPOSITIONIK-RESPONSE
KINEMATICS_MSGS-SRV:GETCONSTRAINTAWAREPOSITIONIK-RESPONSE is not
of type SYMBOL". Due to this, the function holding the actual
functionality is renamed to `ik->trajectory-ex' and two separate
adapter functions `ik->trajectory' are created, each stating the
type of message they accept in the parameters list. They in turn call the actual `ik->trajectory-ex' function, relieving it from it's type checking duties.