Restructured arm and handle reasoning to be more accessible from prolog
The functions used when reasoning about the nearest handle and the most appropriate arm side are split up and made more accessible when used from prolog. A list of handles, the arms from which they are reachable and their respective distance is now generated and used to determine the best choice. Also, fixed indentation and doc strings at a few spots.