Moved all handle and arm side reasoning into prolog
The lisp functions for determining the nearest handle and arm side is still coded in lisp but is being called from prolog. The `best-grasp' predicate now contains all necessary steps for calculating the nearest handle instead of just blindly calling one function that returns the nearest one. Code cleanup and a bit of refactoring is still to be done. Also, added documentation and comments.