Changed a debug printout and removed an unused, old function
The function to calculate the best arms and handles for a certain situations was removed due to the fact that it was reimplemented in a more prolog friendly manner. This old function was a leftover piece and is now removed. Also, changed a debug printout to use `ros-info' instead of `format'.