Added prolog predicates and lisp functions for multi handle/arm grasp evaluation
To decide which arm to use to grasp with which handle in multi-arm/-handle situations, functions for calculating the optimal gripper distance from the solution were implemented. The PR2 now determines which handle is grasped with which arm based on the respective distance between them. This way, objects can be grasped with an arbitrary number of arms in the most distance-optimal way. A future extension to this is the decision of arm-/handle-choice based on the stability of the handle-grasp.