Added `(par-synced' macro and made `grab-handled-object' use multiple arms
The macro `(par-synced' macro was added to execute a sequence of blocks of parallel actions. Meaning:

(par-synced (A B) (C D E))

A and B are executed in parallel. When they are done, C, D and E are executed in parallel. This is applied to the `grab-handled-object' method to make it grab multi-handled objects and use several grippers in parallel. The synchronized parallelization is necessary here due to the fact that each gripper has to be in pose / opened/closed before the next step can be started (otherwise, objects might slip, trajectories are not in sync anymore, etc.). Also, a `ros-info' was added to inform the user that the optimal grasp arm/handle combination is currently being calculated (this process can take a few seconds and led to confusion sometimes).