Made the holding-arms predicate only rely on the belief state
The belief state holds all information necessary to identify the arms that currently hold an object in question. Though, the information is still stored in the designator and should be used accordingly at appropriate places where a query to the belief state would be too `expensive'. Since at the moment, only single-arm grasps are stored in the designator, this solution is commented out and should be readded after refactoring the appropriate functions updating the designator.