Removed the constraint-awareness for the actual grasp for handled objects
Since the grippers need to get real close to the object when executing the actual grasping motion (after the pregrasp), the constraint-awareness needs to be switched off here. Otherwise, the collision environment will signal an invalid IK solution and will break the whole grasp. As we are already in pregrasp when this happens, we assume that no obstacles are between the grippers and the handles to grasp.