Added parking function for two arms
When an object is grasped with two arms, `park-grasped-object-with-two-arms' is called. Wiring is already finished. The object is placed in front of the torso_lift_link. At the moment, the same z-level for all grasps is assumed. Should this not be the case in the future, the function has to be adapted.