Removed tf frame name workaround and fixed a wrongly rotated handle grasp pose
The handle grasp pose which determines whether a grasp is valid was based on a wrong rotation. As handles on objects are defined as pointing outwards of the object (into the approach direction), this rotation was not inverted when determining whether a grasp for that handle is do-able or not. This rotation is added now.
Also, several transformations into "/torso_lift_link" were done. These are now removed and replaced by transforms to "/map" which makes much, much more sense.