Moved grasping into separate file, added sensors
The grasping-related code is now found in the `grasping.lisp' file rather than the `handled-objects.lisp' file. The reason is that the grasping code was refactored to spread over objects with as well as without handles.
Also, added fingertip sensors handling code. The function `(pressure-fingertip side finger)' now gives a pseudo-filtered value of the current requested sensor. The function `(zero-fingertip-pressure side)' must called before using the sensors. When opening a gripper, the respective side's zero point is set automatically.