Added package `visibility_costmap'
The visibility-costmap package includes functions and predicates to resolve location designators such as `(visible-for robot)', `(invisible-for robot)', and `(occluded-by ,obj)' and generates costmaps over a certain region in which these descriptions hold.

It builds on top of semantic_map_costmap and bullet_reasoning to take the semantic map into account and use bullet's visibility-reasoning capabilities.