Refactored put-down code for grasped objects
Major overhaul of put-down code and fix of a bug in execute-move-arm-pose (the verification of successful arm moves tended to fail for movements that took long). Grasp and put-down is now collision aware, using the motion planners and the collision environment. When placing objects, they are attached as collision objects to the robot's gripper(s). The selection of collision shapes could be refined, though. Also, executing arm-movements with the motion planners has to feature two different (ros) namespaces when executing multi-hand movements. This is still to be done.