Changed the object-in-hand goal so that it retries grasping several times
Instead of re-perceiving the object to grasp instantly when grasping didn't work, try to grasp it from another angle again. This happens three times before it is re-perceived. This way, the PR2 gets more chances to actually get the object into it's gripper(s) without instantly moving far away to get a better view.

Also, the location described as `(to execute)' the grasp actions is replaced by the `(to reach)' description. This expresses the intent of the location very clearly.