Refactored grasping-related code to work with containers and MoveIt!
Incorporating the MoveIt! related functionality now solves movement of the PR2 arms. Also, container related grasping (and grasping in general) is refactored to

a) include more information (relative and absolute handles are handed through the functions)
b) Opening of containers was introduced (up to a given angle, this still has to be refined)