Added macro and adjacent infrastructure for defining policies during plans
In order to define a policy for use while plan performance, three code blocks and a list of parameters must be supplied:

(define-policy object-moved (object)
  (:check (format t "Testing for movement of object ~a~%" object)
          (do-object-checking-here)
  (:clean-up (format t "Cleaning up~%")
  	     (do-clean-up-here))
  (:recover (format t "Recovering~%")
  	    (do-recovery-here))))

During a plan, call it like this:

(with-policy object-moved (?obj)
  (do-plan-actions-with-policy-awareness))

The parameter `object' defined during definition of the policy is replaced by `?obj' during the call of `with-policy' and is available as `object' to all code blocks of the policy (currently `check', `clean-uo', `recover').

`check' is called whenever the policy is checked for triggering.
`clean-up' is called during the clean-up-phase of unwind-protect (so it is called either way)
`recover' is called when the policy gets triggered before the main body of `with-policy' is completed.

If the main body is completed without the policy getting triggered, only `clean-up' is called and should be treated like that. `with-policy' returns the return value of the main body. If a policy is referred to that was not defined before (i.e. does not exist), a plan failure of type `policy-not-found' is signalled.