Only consider *running* process modules when performing designators
Up to now, loaded process modules are considered for designator resolution even if they are not running (`(with-process-modules-running ...)` and friends). This leads to unspecified behavior, and (most prominently) a failure claiming that the designator could not be resolved.

The most difficult situation that this resolves comes up when having loaded, but not running, several process modules that respond to the same designators.