Added support for documentation strings given right before actual body code
`with-policy' now supports being defined with an optional documentation string, just as in `defun' and friends. So defining a policy would now look like this:

(define-policy my-policy (max-num match-num)
  "This is an example policy."
  (:init (format t "Initializing policy~%")
         t)
 ...)

At the same time, got rid of the `:description' keyword in the properties list (which served as entry point for policy documentation before).