Added generic utilities package with minium dependencies
The ubiquitous utilities package is supposed to act as middle layer between high level (very abstract) components and low level (very specific) routines. The main motivation for this are the hard-coded calls to TF all over the place in cram_highlevel. Since we want to replace this with TF2 (and not just replace the hard-coded stuff, but move on to something more generic in case similar situations come up in the future), this package acts as a middle layer.

Other components (as loaded) can register utility functions for specific symbols (e.g. ':transform-pose') while these functions can then be called from any other component using the ubiquitous utilities package.