Don't fall back to one carry handle after valid carry-handles desig property
When evaluating the prolog predicate for the action `(to grasp)`, a fallback value for the number of handles to use for carrying (`carry-handles`) is equal to the amount 1. The former implementation worked well then no such property was present in the respective object designator, but lead to backtracking in the prolog rules when one was present, and the grasp ultimately failed (thus, retrying the whole grasp, but with only one carry-handle, despite the property possibly being set differently in the designator). This is fixed now.