Put-down code restructured, started to resolve single-hand hack
Like grasping with multiple hands, putting down objects with the same amount of hand in their respective (object-)relative poses needs to be restructured in order to support movement of multiple robot links in parallel. The presented changes introduce a new structure for the `put-down` action handler that strive to serve this purpose.