[plan_lib] commented out TF-callback related code for AT-LOCATION:
it was used to check the location constraints of AT-LOCATION form
in case the robot drives too far away while executing the task
and AT-LOCATION constraints would not hold anymore.
In our plans we use simple tasks per AT-LOCATION call,
so this should not happend.
TF callbacks were too low-level for the plan library, so now they're gone.