Setting `:timestamp` to `0.0` for `on-event` `robot-state-changed`
The timestamp setting in the robot-state-changed decides which timestamp is sent to tf for `lookup-transform`. It defaults to the current time, which can result in tf running into its timeout when waiting for transforms. Manually setting it to `0.0` results in tf using the latest transform available (which usually works).

If we ever want to use tf's time travel feature, this becomes problematic. But for the time being, this is working perfectly.