new function for axis spawning and got rid of dependency on json_prolog
The function is in the cram_bullet_reasoning and its called "add-vis-axis-obect 'name".

You can call it with: (btr:add-vis-axis-object 'mug1) optional you can also set the length of the axis as optional parameter. You can clear everthing with the function: (btr:clear-costmap-vis-object). Colors are: red for x, green for y and blue for z.