[desig] action desigs should not be equated to their resolved actions:
because the resulting action designator is not effective, calling current-desig will always give an unresolved desig.
In general, for getting the latest effective designator, one should call desig:newest-effective-desig.
It is bad in any case to equate them, as resolving an fully specified action designator is not really implemented at the moment...