[visibility-cm] commented out collision check in visibility check:
there's more to collisions than just checking if robot's not in collision with attached
objects. What if he's in collision with objects that the attached objects are in collision with. What if there is no attachment between these objects but they actually belong to the same object as in case of the assembly plane. Blabla...