moved cl_quadtree contents into cram_location_costmap:
cl_quadtree couldn't stay in cram_3rdparty because it depended on a ROS package cl_transforms,
it also couldn't go to cram_math, as it depended on ROS,
it also couldn't go to roslisp_common, as quadtrees are not a part of the ROS API.
So it ended up inside cram_location_costmap, for which it was actually developed.
If the author (@tlipps) intended it to be a standalone library, he shouldn't have depended on an obscure ROS Lisp package cl_transforms.