[fd-plans] commented out test deps and make bowl weight 0:
test dependencies in the package.xml might make unnecessary things get installed and create dependency loops. As long as the deps are in the ASDF file, it's fine.

Concerning the bowl, it seemed to keep going through surfaces upon simulation :/ Giving it 0 weight made gravity not affect it. I wonder why the mesh is so bad :(