[mpp-plans] do not ignore manipulation failures of PUTTING action:
when experimenting on the real robot, at least dropping the object can be done as last resort, but in simulation that makes only a bigger mess than handling the failure. On the real robot it also looks very clumsy and noisy and all that, so handle the failure and retry instead.